refactor version command
